Login

+++ src/app/posts/[postId]/page.tsx
@@ -5,9 +5,8 @@ import { usePosts } from "@/app/context";
 import Image from "next/image";
 import { Button } from "@/components/ui/button";
 import { Trash2, Pencil } from "lucide-react";
-import { BlurImage } from "@/components/gallery";
 import { cn } from "@/lib/utils";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import {
   Dialog,
   DialogTrigger,
@@ -17,23 +16,52 @@ import {
   DialogFooter,
   DialogDescription,
 } from "@/components/ui/dialog";
-import Link from "next/link";
 import EditPostDialog from "@/components/edit-post-dialog";
+import { Post, PostWithUser, UserPost } from "@/types";
+import { useUser } from "@/app/context/usercontext";
 
 export default function SinglePostPage() {
   const params = useParams();
   const router = useRouter();
-  const { posts, deletePost } = usePosts();
+  const { user } = useUser();
+  const { posts, deletePost, getPost } = usePosts();
 
-  const post = posts.find((p) => p.id.toString() === params.postId);
+  const [post, setPost] = useState<PostWithUser | null>(null);
   const [isLoading, setLoading] = useState(true);
   const [open, setOpen] = useState(false);
+
+  const { postId } = useParams();
+
+  const fetchPost = async () => {
+    setLoading(true); // Ensure loading is true before fetching.
+    const fetched = await getPost(postId as string);
+    if (!fetched) {
+      setPost(null); // Post not found.
+    } else {
+      fetched.date = new Date(fetched.date); // Convert date string to Date object.
+      setPost(fetched); // Update state with fetched post.
+    }
+    setLoading(false); // Turn off loading when data is fetched.
+  };
+
+  useEffect(() => {
+    if (!postId) return;
+    fetchPost();
+  }, []);
+
+  if (isLoading) {
+    return <p className="text-center text-gray-500 text-xl">Loading...</p>;
+  }
 
   if (!post) {
     return <p className="text-center text-red-500 text-xl">Post not found.</p>;
   }
 
   const handleDelete = () => {
+    if (post.user_id !== user?.id) {
+      alert("You are not authorized to delete this post.");
+      return;
+    }
     deletePost(post.id);
     setOpen(false);
     router.push("/"); // Redirect after deleting
@@ -41,18 +69,20 @@ export default function SinglePostPage() {
 
   return (
     <div className="py-5 flex flex-col justify-center items-center">
-      <div className="absolute top-4 right-4 flex space-x-2">
-        <EditPostDialog post={post} />
-        <Button
-          data-testid="delete-post"
-          onClick={() => setOpen(true)}
-          size="icon"
-          className="bg-[#181832] hover:bg-pink-500 text-white rounded-full"
-          variant="destructive"
-        >
-          <Trash2 size={18} />
-        </Button>
-      </div>
+      {post.user_id === user?.id && (
+        <div className="absolute top-4 right-4 flex space-x-2">
+          <EditPostDialog post={post} />
+          <Button
+            data-testid="delete-post"
+            onClick={() => setOpen(true)}
+            size="icon"
+            className="bg-[#181832] hover:bg-pink-500 text-white rounded-full"
+            variant="destructive"
+          >
+            <Trash2 size={18} />
+          </Button>
+        </div>
+      )}
       <div className="relative priority w-full rounded-lg max-w-5xl mx-auto aspect-[16/9]">
         <Image
           alt=""
@@ -73,6 +103,7 @@ export default function SinglePostPage() {
         <h1 className="text-4xl font-bold">{post.title}</h1>
         <p className="text-gray-600 text-lg">Type: {post.type}</p>
         <p className="text-gray-700">{post.subject}</p>
+        <p className="text-gray-700">{post.username}</p>
         <p className="text-gray-500 text-sm mt-2">
           {post.date.toLocaleDateString("en-GB")}
         </p>
